Refactor: Move Query AST to relvar-core
This commit moves the Query AST and execution logic from `relvar/src/experimental/query.rs` to `relvar-core/src/query/`. This improves architectural boundaries by placing the query logic in the core crate and reducing type duplication.

- Add `serde` support to `Aggregation` and `AggregationFn` in `relvar-core`.
- Create `relvar-core/src/query` module with `ast.rs` and `error.rs`.
- Implement `Query::execute` in `relvar-core`.
- Deprecate `relvar/src/experimental/query.rs` and re-export core types.
