+++ relvar-core/src/query/error.rs
@@ -1,0 +1,21 @@
+//! Errors that can occur during query execution.
+
+use crate::constraints::ExpressionError;
+use crate::error::DatabaseError;
+use thiserror::Error;
+
+/// Errors that can occur during query execution.
+#[derive(Debug, Error)]
+pub enum QueryError {
+    /// Error from the database engine.
+    #[error("Database error: {0}")]
+    Database(#[from] DatabaseError),
+
+    /// Error evaluating a constraint expression.
+    #[error("Constraint evaluation error: {0}")]
+    Constraint(#[from] ExpressionError),
+
+    /// Error in relational algebra operation.
+    #[error("Algebra error: {0}")]
+    Algebra(String),
+}
